requirements.txt
google-genai
python-dotenv
neonize

